rename katana to providers
