+++ account_sdk/src/tests/mod.rs
@@ -2,7 +2,6 @@ use std::net::TcpListener;
 
 use starknet::{core::types::FieldElement, signers::SigningKey};
 
-mod contract_test;
 mod katana_test;
 mod new_deploy_test;
 mod one_file_all;
